refactor!: adopt the sql.dialect/sql.model coordinates from the jdbc.db split
The dialect framework and the SQL vocabulary moved to the
org.eclipse.daanse.sql repository (plain file copies; history in
jdbc.db). Mechanical rename: imports/deps/bndrun BSNs
org.eclipse.daanse.jdbc.db.dialect.* -> org.eclipse.daanse.sql.dialect.*,
org.eclipse.daanse.jdbc.db.api.type -> org.eclipse.daanse.sql.model.type;
the test bndruns additionally run the sql.model bundle.

Also: IsDateCalcTest pins Locale.US — its fixture strings are US-locale
forms and IsDate parses with the default locale (failed under en_GB).
