+++ api/pom.xml
@@ -36,7 +36,7 @@
 		</dependency>
 		<dependency>
 			<groupId>org.eclipse.daanse</groupId>
-			<artifactId>org.eclipse.daanse.jdbc.db.dialect.api</artifactId>
+			<artifactId>org.eclipse.daanse.sql.dialect.api</artifactId>
 			<version>0.0.1-SNAPSHOT</version>
 		</dependency>
 		<dependency>
